+++ crates/lib3h/tests/test_suites/mod.rs
@@ -2,3 +2,4 @@ pub mod three_basic;
 pub mod two_basic;
 pub mod two_get_lists;
 pub mod two_spaces;
+pub mod two_connection;
